Update SpecSuite.cls
The idea is to have a line, telling us how many times have we tried to add tests to the tool. In some cases, it may execute only the tests it can and thus the final line should look with this:

= PASS (48 of 48 passed) = 18.08.2016 15:24:53 =========================
Total tests:49

Then we realize that something was worng.